// src/types/utools.d.ts
export interface CallbackListItem {
  title: string
  description: string
  icon?: string
  url?: string

  [propName: string]: any
}